Auto virtical split size change; other improvements
This commit makes things just move a little more quickly. I was getting
some strange delays with my spaces due to using space as my leader, but
this seems to have corrected that.

I also finally figured out a way for my splits to autoadjust to a size I
like, which allows for me to use larger font sizes.